only change data if loged in

+++ php/actions/edit_game_action.php
@@ -8,36 +8,41 @@ $listDAO = new PlayerListDAO();
 require_once "../../db/game_dao.php";
 $gameDAO = new GameDAO();
 
-$gameid = $gameDAO->getGameByName($_SESSION['gamechoice'])->gameid;
-$userid = $_SESSION['userid'];
+$isLoggedIn = $_SESSION['userid']>-1;
+
+if($isLoggedIn){
+
+    $gameid = $gameDAO->getGameByName($_SESSION['gamechoice'])->gameid;
+    $userid = $_SESSION['userid'];
 
 
-if(isset($_POST['deletegame'])){
-    $errorcode = $listDAO->deletePlayer($gameid, $userid);
+    if(isset($_POST['deletegame'])){
+        $errorcode = $listDAO->deletePlayer($gameid, $userid);
 
-}elseif(isset($_POST['savegame'])){   
+    }elseif(isset($_POST['savegame'])){   
 
-    $rank = $_POST['rank'];
-    $roles =[];
-    if(isset($_POST['role']))
-        $roles = $_POST['role'];
-    if(isset($_POST['visible'])) {
-        $status = 'active';
-    } else{
-        $status = 'inactive';
-    }
+        $rank = $_POST['rank'];
+        $roles =[];
+        if(isset($_POST['role']))
+            $roles = $_POST['role'];
+        if(isset($_POST['visible'])) {
+            $status = 'active';
+        } else{
+            $status = 'inactive';
+        }
 
-    /* Check if player is already in playerlist for specific game*/
-    switch($listDAO->alreadyIncluded($gameid, $userid)){
-        case false:
-            $listDAO->addPlayer($gameid, $userid, $rank, $roles, $status);
-            break;
-        case true:
-            $listDAO->updatePlayer($gameid, $userid, $rank, $roles, $status);
-            break;
-        default:
-            // TODO Errormessage
-            break;
+        /* Check if player is already in playerlist for specific game*/
+        switch($listDAO->alreadyIncluded($gameid, $userid)){
+            case false:
+                $listDAO->addPlayer($gameid, $userid, $rank, $roles, $status);
+                break;
+            case true:
+                $listDAO->updatePlayer($gameid, $userid, $rank, $roles, $status);
+                break;
+            default:
+                // TODO Errormessage
+                break;
+        }
     }
 }
 header('Location: ../../change_profile.php');
